+++ indie-box/vendor_perl/IndieBox/AppConfigurationItems/File.pm
@@ -50,12 +50,14 @@ sub new {
 }
 
 ##
-# Install this item
+# Install this item, or check that it is installable.
+# $doIt: if 1, install; if 0, only check
 # $defaultFromDir: the directory to which "source" paths are relative to
 # $defaultToDir: the directory to which "destination" paths are relative to
 # $config: the Configuration object that knows about symbolic names and variables
-sub install {
+sub installOrCheck {
     my $self           = shift;
+    my $doIt           = shift;
     my $defaultFromDir = shift;
     my $defaultToDir   = shift;
     my $config         = shift;
@@ -100,8 +102,10 @@ sub install {
 
             my $contentToSave = $templateProcessor->process( $content, $config, $sourceOrTemplate );
 
-            unless( saveFile( $toName, $contentToSave, $mode, $uname, $gname )) {
-                error( 'Writing file failed:', $toName );
+            if( $doIt ) {
+                unless( saveFile( $toName, $contentToSave, $mode, $uname, $gname )) {
+                    error( 'Writing file failed:', $toName );
+                }
             }
 
         } else {
@@ -111,12 +115,14 @@ sub install {
 }
 
 ##
-# Uninstall this item
+# Uninstall this item, or check that it is uninstallable.
+# $doIt: if 1, uninstall; if 0, only check
 # $defaultFromDir: the directory to which "source" paths are relative to
 # $defaultToDir: the directory to which "destination" paths are relative to
 # $config: the Configuration object that knows about symbolic names and variables
-sub uninstall {
+sub uninstallOrCheck {
     my $self           = shift;
+    my $doIt           = shift;
     my $defaultFromDir = shift;
     my $defaultToDir   = shift;
     my $config         = shift;
@@ -133,7 +139,9 @@ sub uninstall {
         unless( $toName =~ m#^/# ) {
             $toName = "$defaultToDir/$toName";
         }
-        IndieBox::Utils::deleteFile( $toName );
+        if( $doIt ) {
+            IndieBox::Utils::deleteFile( $toName );
+        }
     }
 }
