+++ indie-box/vendor_perl/IndieBox/AppConfigurationItems/Symlink.pm
@@ -49,12 +49,14 @@ sub new {
 }
 
 ##
-# Install this item
+# Install this item, or check that it is installable.
+# $doIt: if 1, install; if 0, only check
 # $defaultFromDir: the directory to which "source" paths are relative to
 # $defaultToDir: the directory to which "destination" paths are relative to
 # $config: the Configuration object that knows about symbolic names and variables
-sub install {
+sub installOrCheck {
     my $self           = shift;
+    my $doIt           = shift;
     my $defaultFromDir = shift;
     my $defaultToDir   = shift;
     my $config         = shift;
@@ -92,7 +94,9 @@ sub install {
         }
         if( -r $fromName ) {
             unless( -e $toName ) {
-                IndieBox::Utils::symlink( $fromName, $toName, $mode, $uname, $gname );
+                if( $doIt ) {
+                    IndieBox::Utils::symlink( $fromName, $toName, $mode, $uname, $gname );
+                }
             } else {
                 error( 'Cannot create symlink:', $toName );
             }
@@ -104,12 +108,14 @@ sub install {
 }
 
 ##
-# Uninstall this item
+# Uninstall this item, or check that it is uninstallable.
+# $doIt: if 1, uninstall; if 0, only check
 # $defaultFromDir: the directory to which "source" paths are relative to
 # $defaultToDir: the directory to which "destination" paths are relative to
 # $config: the Configuration object that knows about symbolic names and variables
-sub uninstall {
+sub uninstallOrCheck {
     my $self           = shift;
+    my $doIt           = shift;
     my $defaultFromDir = shift;
     my $defaultToDir   = shift;
     my $config         = shift;
@@ -126,7 +132,9 @@ sub uninstall {
         unless( $toName =~ m#^/# ) {
             $toName = "$defaultToDir/$toName";
         }
-        IndieBox::Utils::deleteFile( $toName );
+        if( $doIt ) {
+            IndieBox::Utils::deleteFile( $toName );
+        }
     }
 }
